array: BUGFIX: find wasn't checking last element
When cl:do(*) is used, all variables are set before the check is made, which happens before the body is called.  This means that we can't set a do variable as the result of accessing the array unless we do all other work in do variables.  The check out happen and drop the last result before the body is called otherwise.