list:split^ functions: fixed from-end and end
End has to be set to one past it's number to make sure that element is included.

Made some changes on how from-end works based on the insight that from-end only has an effect when count is also present, and in that case we just take the last "count" results.  This means that we can just run as normal and ignore count until we have all results, then apply count before reversing and returning them.